Unify AllowedAddressConverter configuration code
